Insert newline at the end of the .netrc file
Due to a regression in the version of curl shipped with macOS Ventura 13.2, there needs to be a newline at the end of the .netrc file for it to be read properly. This is fixed in the next version of curl, but we have no guarantees as to when that will be included.